Move `RestoreSink` to `fs-sink.cc`
